requirements.txt
dj-static==0.0.6
dj-database-url==0.5.0
Django==2.2.7
django_adminlte==1.2
django-crispy-forms==1.8.1
django-cruds-adminlte==0.0.16
django-dynamic-formsets==0.0.8
django-debug-toolbar==2.1
django-decouple==2.1
django-extensions==2.2.1
django-pandas==0.6.1
django-select2==7.2.0
django-widget-tweaks==1.4.5
gunicorn==20.0.4
pandas==0.25.3
psycopg2==2.8.4
unipath==1.1
